Refactor some code duplication
